+++ tests/structures/testResourceIdentifier.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock
 
 from tests.testCase import testCase
 
-from exonetapi.structures import ResourceIdentifier
-from exonetapi.structures.Resource import Resource
+from exonetapi.structures import ApiResourceIdentifier
+from exonetapi.structures.ApiResource import ApiResource
 from exonetapi.structures.Relationship import Relationship
 from exonetapi.structures.Relation import Relation
 from exonetapi import create_resource
@@ -71,9 +71,9 @@ class testResourceIdentifier(testCase):
         resource.reset_changed_relations()
 
         resource.relationship('object', {'data': {'type': 'this', 'id': 'that'}})
-        resource.relationship('resource', Resource('this', 'that'))
-        resource.relationship('resource_identifier', ResourceIdentifier('this', 'that'))
-        resource.relationship('list', [ResourceIdentifier('this', 'that')])
+        resource.relationship('resource', ApiResource('this', 'that'))
+        resource.relationship('resource_identifier', ApiResourceIdentifier('this', 'that'))
+        resource.relationship('list', [ApiResourceIdentifier('this', 'that')])
 
         self.assertEqual(
             resource.get_json_changed_relationships(),
@@ -110,7 +110,7 @@ class testResourceIdentifier(testCase):
         )
 
     def test_to_json(self):
-        resource = Resource({
+        resource = ApiResource({
             'type': 'fake',
             'id': 'FakeID',
         })
@@ -151,7 +151,7 @@ class testResourceIdentifier(testCase):
     @mock.patch('exonetapi.RequestBuilder.get')
     def test_post(self, mock_requestbuilder_get, mock_requestbuilder_init):
         mock_requestbuilder_get.get = MagicMock(return_value=None)
-        Resource({'type': 'fake', 'id': 'FakeID'}).get()
+        ApiResource({'type': 'fake', 'id': 'FakeID'}).get()
         mock_requestbuilder_get.assert_called_with('FakeID')
         mock_requestbuilder_init.assert_called_with('fake')